refactor(barbrief): use shared 10-heading validator, matching strategy
barbrief previously validated case facts with a looser substring check (a
heading word appearing anywhere) and was silent on failure; strategy used an
anchored per-line regex (each heading on its own line, case-insensitive,
punctuation-tolerant) and reported missing headings. Make them identical by
extracting strategy's validator to a shared module and using it in both.

- litassist/utils/case_facts.py: new home for validate_case_facts_format.
- strategy/validators.py: re-exports it (existing importers/patches unchanged);
  dropped the now-unused click import.
- barbrief/core.py: uses the shared validator; barbrief/validator.py deleted.
- tests: removed the redundant barbrief validator unit tests (strategy's
  comprehensive tests cover the shared function); retargeted the barbrief patch
  in test_command_parameters; updated the barbrief command fixtures to put each
  heading on its own line (the stricter rule the substring check had masked).

Behavioural change: barbrief now rejects case facts whose headings are not each
on their own line, matching strategy.
