+++ CHANGELOG.md
@@ -28,6 +28,9 @@ Historical dated sections preserve the model names that were current when those 
 
 ### Changed
 
+#### May 2026: barbrief 10-heading validation matches strategy (shared validator)
+- The 10-heading case-facts check is now a single shared function, `validate_case_facts_format` in `litassist/utils/case_facts.py`, used by both `strategy` and `barbrief` (re-exported from `litassist/commands/strategy/validators.py` for existing importers). `barbrief` previously used a looser substring check (`barbrief/validator.py`, now removed) that accepted a heading appearing anywhere in the text. It now uses the same anchored per-line regex as `strategy`: each of the ten headings must be on its own line (case-insensitive, punctuation-tolerant), and missing headings are listed. Behavioural change: `barbrief` rejects case facts whose headings are not each on their own line, matching `strategy`/`extractfacts` output.
+
 #### May 2026: parameter translation aligned with recent GPT-5 / o3 / Grok / Gemini changes
 - `litassist/llm/parameter_handler.py` and `model_profiles.py` now map generated requests to each model's current accepted parameters. GPT-5.5 gained the `xhigh` reasoning tier, so the universal `thinking_effort: max` maps to `xhigh` for `gpt5.5` (its ceiling); older GPT-5 variants and the o-series still cap at `high`. `verbosity` is now skipped for Grok 4.x (alongside the o-series) because those models do not accept it, and the `xai` profile was trimmed to grok-4.20's supported set -- `frequency_penalty`/`presence_penalty`/`verbosity`/`stop` dropped, while `min_p`/`top_a`/`repetition_penalty` continue to ride `extra_body` as best-effort. The o-series (o3-pro) already mapped correctly and is unchanged; Gemini now also skips `verbosity` (which `gemini-3.5-flash` does not accept and the `lookup` configs were setting). Both are pinned by tests.
 - A directly-supplied `reasoning.effort` (a caller passing `reasoning={"effort": ...}` rather than `thinking_effort`) is now normalised through the same per-family mapping, so it cannot bypass the effort caps (e.g. `xhigh`/`max` reaching o3 or sonnet); an `effort` of `none` or `null` drops the reasoning object entirely (it is not coerced to medium).
